// src/enums/route.ts
const enum ROUTE {
    CLASSIFY_MANAGE = "/main/classifyManage",
    INFO_MANAGE = "/main/infoManage",
    CLASSIFY_NEW = "main/newClassify",
    CLASSIFY_REVISE = "/main/reviseClassify",
    CLASSIFY_VIEW = "/main/viewClassify",
    INFO_NEW = "/main/newInfo",
    INFO_REVISE = "/main/reviseInfo",
    INFO_VIEW = "/main/viewInfo"
}

export default ROUTE;